requirements.txt
aniso8601==9.0.1
anyio==4.0.0
apispec==6.3.0
apispec-webframeworks==0.5.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.7.0
cachelib==0.9.0
certifi==2023.11.17
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
flasgger==0.9.7.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
flask-restful-swagger==0.20.2
flask-restful-swagger-2==0.35
flask-restful-swagger-3==0.5.1
flask-restplus==0.13.0
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
greenlet==3.0.1
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-resources==6.1.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mistune==3.0.2
openapi-schema-validator==0.2.3
openapi-spec-validator==0.4.0
packaging==23.2
pika==1.3.2
prance==0.21.8.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyrsistent==0.20.0
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
semver==2.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
validators==0.22.0
Werkzeug==3.0.1
WTForms==3.1.1
